Enhanced documentation + separated support class #3945
